requirements.txt
# Python dependencies for Tape 3D Generator
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0

# AI Training dependencies
openai>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
jupyter>=1.0.0